refactor: Ladder 클래스에 simulate 기능 추가
