+++ src/test/java/nextstep/ladder/domain/LadderTest.java
@@ -11,6 +11,7 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 class LadderTest {
@@ -18,17 +19,28 @@ class LadderTest {
     @DisplayName("Ladder 인스턴스 만들기")
     @Test
     public void testConstructor() {
-        List<LadderUser> ladderUsers = List.of(
-            new LadderUser("pobi", 0),
-            new LadderUser("honux", 1)
+        LadderUsers ladderUsers = new LadderUsers(
+            List.of(new LadderUser("pobi", 0), new LadderUser("honux", 1))
         );
-        List<LadderPrize> ladderPrizes = List.of(
-            new LadderPrize("꽝"),
-            new LadderPrize("5000")
+        LadderPrizes ladderPrizes = new LadderPrizes(
+            List.of(new LadderPrize("꽝"), new LadderPrize("5000"))
         );
 
-        LadderUsers users = new LadderUsers(ladderUsers);
-        LadderPrizes prizes = new LadderPrizes(ladderPrizes);
-        assertDoesNotThrow(() -> new Ladder(new LadderConfig(users, prizes), new Dimension(3), new LineFactory()));
+        assertDoesNotThrow(() -> new Ladder(ladderUsers, ladderPrizes, new Dimension(3), new LineFactory()));
+    }
+
+    @DisplayName("LadderC 인스턴스 생성 - 유저와 결과 수가 맞지 않으면 예외 반환")
+    @Test
+    public void testConstructor_throwException() {
+        LadderUsers ladderUsers = new LadderUsers(
+            List.of(new LadderUser("pobi", 0), new LadderUser("honux", 1))
+        );
+        LadderPrizes ladderPrizes = new LadderPrizes(List.of(new LadderPrize("1000")));
+
+        assertThatThrownBy(() -> new Ladder(ladderUsers, ladderPrizes, new Dimension(3), new LineFactory()))
+            .isInstanceOf(IllegalArgumentException.class)
+            .hasMessageContaining("유저와 결과 수가 맞지 않습니다.");
     }
 }
+
+
